Homework 3 Changes
The main (index) page now has a link to a form page.  This page contains two
forms, one which sends a GET request, and one which sends a POST request.
They both redirect to a submit page, which displays the user-entered values.

server.py also has a lot of refactoring, so now there is a function for every
different path, each with params "conn" and "args".  This is for "future
purposes".

tests_server.py has renamed tests to match the new functions, and additional
tests to make sure the form and submit pages work correctly.  Most of the
changes are just removing spaces, since the responses inside server.py are no
longer nested in if statements, so they have fewer spaces out front.

ChangeLog was updated for Homework 3