Homework 6 Changes
Lots of changes to app.py so that it can serve images and text files.
Each handle_***_get function now renders its own template or file,
instead of it happening all in one place in simple_app.
Some refactoring of header parsing to make it more readable.

Modified submit.html to fix the issue with only the first letter
appearing (which made my HW5 tests fail).

Modified test_server.py so it doesn't spit out hundreds of lines of
encoded image.  It doesn't really test whether the image is valid at
the moment, but I plan to ask about that.

Modified server.py because image and file responses don't have a list
of things in "ret", so that for loop was problematic.

city_all.txt is an example file used in the CSE 231 lab I led today.
http://www.cse.msu.edu/~cse231/Labs/Lab07/city_all.txt

galaxy.jpg was listed as free for non-commercial use on Wikipedia.
http://commons.wikimedia.org/wiki/File:Blackeyegalaxy.jpg